pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roic-builder"
version = "0.1.0"
description = "ROIC Builder: ingestion, synthesis, and slide automation pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "ROIC Builder Team" }]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
roic-builder = "roic_builder.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
